Added correlation features, removed numpy dependance.
Information dictionary now stores original data array if appropriate,
this can be used to calculate the correlation and covariance of two
measurement objects. All functions now behave as derivative or Monte
Carlo. Numpy is now only required for the Monte Carlo method, if not
installed, the module will default to the derivative method and print
out a statement to the user. Groundwork is also set for automatic
differentiation.