Constructor for MeasuredValue
A basic implementation of the constructor wrapper of MeasuredValue.

MeasuredValue is renamed from Measurement. The wrapper constructor
is named Measurement. It will check the inputs, and return a basic
MeasuredValue object for single measurements, or for an array of
repeated measurements of the same value, a RepeeatedlyMeasuredValue
object will be returned

This commit also organizes some import statements